LayoutTests/fast/custom-elements/reactions-for-webkit-extensions.html

<!DOCTYPE html>
<html>
<head>
<title>Custom Elements: CEReactions on HTMLElement and Attr interfaces</title>
<meta name="author" title="Ryosuke Niwa" href="mailto:rniwa@webkit.org">
<meta name="assert" content="webkitdropzone of HTMLElement and Attr interface must have CEReactions">
<meta name="help" content="https://dom.spec.whatwg.org/#node">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../imported/w3c/web-platform-tests/custom-elements/resources/custom-elements-helpers.js"></script>
<script src="../../imported/w3c/web-platform-tests/custom-elements/reactions/resources/reactions.js"></script>
</head>
<body>
<div id="log"></div>
<script>

testReflectAttribute('webkitdropzone', 'webkitdropzone', 'copy', 'move', 'dropzone on HTMLElement');

testAttributeMutator(function (element, name, value) {
    element.attributes[name].firstChild.data = value;
}, 'data on CharacterData');

test(function () {
    var element = define_new_custom_element(['title']);
    var instance = document.createElement(element.name);
    instance.setAttribute('title', 'hello');
    assert_array_equals(element.takeLog().types(), ['constructed', 'attributeChanged']);
    instance.attributes.title.firstChild.appendData(' world');
    var logEntries = element.takeLog();
    assert_array_equals(logEntries.types(), ['attributeChanged']);
    assert_attribute_log_entry(logEntries.last(), {name: 'title', oldValue: 'hello', newValue: 'hello world', namespace: null});
}, 'appendData on CharacterData must enqueue an attributeChangedCallback when mutating an existing attribute');

test(function () {
    var element = define_new_custom_element(['title']);
    var instance = document.createElement(element.name);
    instance.setAttribute('title', 'foo');
    assert_array_equals(element.takeLog().types(), ['constructed', 'attributeChanged']);
    instance.attributes.title.firstChild.insertData(2, 'bar');
    var logEntries = element.takeLog();
    assert_array_equals(logEntries.types(), ['attributeChanged']);
    assert_attribute_log_entry(logEntries.last(), {name: 'title', oldValue: 'foo', newValue: 'fobaro', namespace: null});
}, 'insertData on CharacterData must enqueue an attributeChangedCallback when mutating an existing attribute');

test(function () {
    var element = define_new_custom_element(['title']);
    var instance = document.createElement(element.name);
    instance.setAttribute('title', 'hello world');
    assert_array_equals(element.takeLog().types(), ['constructed', 'attributeChanged']);
    instance.attributes.title.firstChild.deleteData(5, 100);
    var logEntries = element.takeLog();
    assert_array_equals(logEntries.types(), ['attributeChanged']);
    assert_attribute_log_entry(logEntries.last(), {name: 'title', oldValue: 'hello world', newValue: 'hello', namespace: null});
}, 'deleteData on CharacterData must enqueue an attributeChangedCallback when mutating an existing attribute');

test(function () {
    var element = define_new_custom_element(['title']);
    var instance = document.createElement(element.name);
    instance.setAttribute('title', 'hello');
    assert_array_equals(element.takeLog().types(), ['constructed', 'attributeChanged']);
    instance.attributes.title.firstChild.replaceData(1, 4, 'i');
    var logEntries = element.takeLog();
    assert_array_equals(logEntries.types(), ['attributeChanged']);
    assert_attribute_log_entry(logEntries.last(), {name: 'title', oldValue: 'hello', newValue: 'hi', namespace: null});
}, 'replaceData on CharacterData must enqueue an attributeChangedCallback when mutating an existing attribute');

test_with_window(function (contentWindow) {
    const element = define_custom_element_in_window(contentWindow, 'custom-element', []);

    const contentDocument = contentWindow.document;
    contentDocument.body.innerHTML = `<select><option></option></select>`;
    const option = contentDocument.querySelector('option');

    const instance = contentDocument.createElement(element.name);
    option.appendChild(instance);
    instance.textContent = 'hello';

    assert_array_equals(element.takeLog().types(), ['constructed', 'connected']);
    assert_equals(option.innerHTML, '<custom-element>hello</custom-element>');

    const select = contentDocument.querySelector('select');
    assert_equals(select.options[0], option);
    select.options.remove(option);
    assert_equals(select.options.length, 0);
    assert_array_equals(element.takeLog().types(), ['disconnected']);
}, 'remove(HTMLOptionElement) on HTMLOptionsCollection must enqueue disconnectedCallback when removing a custom element');

test_with_window(function (contentWindow) {
    const element = define_custom_element_in_window(contentWindow, 'custom-element', []);

    const contentDocument = contentWindow.document;
    contentDocument.body.innerHTML = `<select><option></option></select>`;
    const option = contentDocument.querySelector('option');

    const instance = contentDocument.createElement(element.name);
    option.appendChild(instance);
    instance.textContent = 'hello';

    assert_array_equals(element.takeLog().types(), ['constructed', 'connected']);
    assert_equals(option.innerHTML, '<custom-element>hello</custom-element>');

    const select = contentDocument.querySelector('select');
    assert_equals(select.options[0], option);
    select.remove(option);
    assert_equals(select.options.length, 0);
    assert_array_equals(element.takeLog().types(), ['disconnected']);
}, 'remove(HTMLOptionElement) on HTMLSelectElement must enqueue disconnectedCallback when removing a custom element');

</script>
</body>
</html>
